requirements.txt
langchain
sentence_transformers
streamlit
qdrant-client
Pypdf2
python-dotenv
openai
torch
huggingface
huggingface_hub 
transformers
InstructorEmbedding